backend/requirements.txt
flask
Flask-SQLAlchemy
torch
Pillow
pandas
python-dotenv
keras
flask_migrate
flask_cors
transformers
tensorflow
selenium
langdetect
googletrans==4.0.0rc1
sentencepiece
gdown
scikit-learn
numpy
requests
requests_oauthlib
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client